refactor: update import paths for project services in folders and trash modules
