+++ services/web/server/src/simcore_service_webserver/folders/_trash_service.py
@@ -17,7 +17,7 @@ from simcore_postgres_database.utils_repos import transaction_context
 from sqlalchemy.ext.asyncio import AsyncConnection
 
 from ..db.plugin import get_asyncpg_engine
-from ..projects._trash_service import trash_project, untrash_project
+from ..projects.projects_trash_service import trash_project, untrash_project
 from ..workspaces.workspaces_service import check_user_workspace_access
 from . import _folders_repository, _folders_service
 from .errors import FolderBatchDeleteError, FolderNotTrashedError
