+++ services/web/server/src/simcore_service_webserver/projects/projects_trash_service.py
@@ -2,12 +2,14 @@ from ._trash_service import (
     batch_delete_trashed_projects_as_admin,
     delete_explicitly_trashed_project,
     list_explicitly_trashed_projects,
+    trash_project,
+    untrash_project,
 )
 
 __all__: tuple[str, ...] = (
     "batch_delete_trashed_projects_as_admin",
     "delete_explicitly_trashed_project",
     "list_explicitly_trashed_projects",
-)
-
-# nopycln: file
+    "trash_project",
+    "untrash_project",
+)  # nopycln: file
